Replacing the rails asset pipeline with webpack
The application should now be run with foreman. You can do so with the following
command:

    foreman start -p 3000

The rails asset pipeline has been disabled and all ruby gems used to manage
JavaScript libraries have been removed from the project's Gemfile.

JavaScript and CSS are now managed by webpack and development of these should
happen in the webpack folder.

Note: The webpack server should be running for the test suite to pass.