Fixing babel and webpack
This adds a configuration file for babel. Previously, it would just pass through
whatever JavaScript it found and pass it directly to the browser. That meant it
was giving the browser pure ES6 in some instances. Now all ES6 should be
translated into ES5.

When that was fixed, the variables defined in application.js were no longer
visible in the web pages themselves. The configuration change makes exports in
the application.js file available on an SDP global variable.