Fixes form content being lost when trying to save.
Previously, if you tried to save a form, but it had an error, like
an invalid OMB control number, you would lose the selected questions
when the page reloaded.

This addresses the issue by saving the form in an AJAX request. If
the save is successful, it redirects the browser to the form show
page. If it fails, it shows the error on the page.

The code is kind of gross in that in mixes jQuery and React/Redux.
This is due to the fact that we are trying to save work on the React
side by using the form helper stuff from Rails. This should all be
refactored so that the form is created and managed by React.