No longer passing routes as props
If a component needs to access Rails routes, it now imports them
directly. Using Rails routes in the components should be avoided anyway
since hitting the server should be done in actions.

Removed the mock_routes as they are no longer needed by the test suite.

Fixed many eslint issues in the tests. Not all, as we have some
commented code in there that we may enable later, but creates some
unused variables.