Profile edits w/o systems or programs in DB
If there are no surveillance systems or programs in the database, the
UI will say that. Makes sure that there are no programs or systems sent
to the server when submitting the request.

This commit also extracts common functionality between the SettingsModal
and the SignUpModal into the new ProfileEditor component.