Fixing issue with draft question editing
If a draft question has otherAllowed set to null, if you select choice
and then never touch the checkbox, otherAllowed is still set to null.
This caused issues for the QuestionDetails component to die when
rendering because it was calling toString() on null. This change uses
the magic of JavaScript falsy values to get around this issue and bring
things in line with other code in the component.