Fix bug breaking redux store shape
This component was mutating objects passed in as props. Those props were
objects from the redux store. Since they are all references to the same
object, mutating them will mutate the redux store, which is bad.

This change creates a clone of the object before mutating it and passing
it along to another component.