Fix display bug on form show
Previous code removed the fetching of all response sets. This introduced
an issue when clicking on the linked response set for a question as it
might not be in the store.

This adds back in the fetching of all response sets. It's not the most
performance conscious solution, but it will work for now and will be
addressed at a later date.