Removing fetch for all questions and response sets
This updated the form show JSON. It now includes the response sets. It
had already included the questions for a form. Middleware now pulls
response sets from forms. With middleware, we can now get the questions
and response sets required by a form from the single form endpoint and
middleware sorts out putting it in the right place in the store.