Fixes bug with creating a question on form edit
If you clicked the "New Question" button on the form edit page, it would
load QuestionModalContainer, which would cause ResponseSetDragWidget to
mount, which triggers a search for response sets. This replaces
searchResults key in the redux store with response sets. After you close
the modal, the old response set search results would hang around and
cause odd behavior.

This fix lets QuestionSearchContainer know that the search results are
in use. When the search results are no longer in use, the
QuestionSearchContainer is notified via props and will execute a search
again to replace searchResults in the redux store with questions.