Fixing bug w/removing response sets from questions
Previously, you could only remove the *last* response set that was added
to a question with a type of choice or open choice. This was because
the remove button is an actual button and will cause the question form
to submit when clicked.

This change catches the click event and then calls preventDefault, which
will prevent the form from submitting.

While this fixes the issue, I still have no idea how you could click the
last item in the list with the previous code and not have it submit the
form.