Reduce webpack bundle size
Two major changes to reduce the size of the JavaScript that is sent to
the browser. First, instead of including all of lodash, including the
individual functions needed. Next, replacing moment with date-fns.
moment does not play nicely with webpack and includes locale files for
everywhere it supports. Webpack can be configured to ignore these, but
the resulting library is still pretty big. date-fns is significantly
smaller and works well with webpack.