Remove double bootstrap
Removing bootstrap-loader. This package had its own webpack entry point
and was loading the JavaScript portions of bootstrap into the
application. It was also loading the bootstrap styles. This was
redundant as the styles were also being incorporated into the
application's scss. This meant that two copies of the bootstrap css
style were being sent to the browser.

This commit removes bootstrap-loader and its webpack entry point.
Loading of the JavaScript part of bootstrap has been moved to
landing.js. References to the webpack entry point have been removed.