Eliminate PropTypes deprecation warning
React is moving the PropTypes functionality out of the main library and
into its own package. This commit reflects that change by adding the
prop-types library to our package json and switching the import
statements to pull PropTypes out of the new package.

This commit also bumps the version of react-bootstrap, since it had a
dependency that would trigger the warning.