More cleanup of ms2p in QuestionEditContainer
This commit gets rid of the need to transform Question.codes into
Question.concepts. If a question is fetched directly from the server,
it will have Question.concepts set correctly. If it comes from search
results, it had Question.codes.

This changes the fetching of search results so that the payload provided
to ADD_ENTITIES_FULFILLED is consistent across server fetches and search
results.