Making item delete in elasicsearch blocking
Prior to this change, when deleting an item through its controller, a
background job would be created to delete the item from the
elasticsearch index. Dependent relationships would also kick off their
own background jobs to update more in the index.

This change pulls the elasticsearch index delete into the request
response cycle for the controller. Updates to the index for dependent
items still take place in a background job.